Remove RollupState which acts as useless proxy
