+++ packages/state/src/index.ts
@@ -3,5 +3,4 @@ export * from './MerkleNode'
 export * from './MerkleTree'
 export * from './MerkleValue'
 export * from './PositionLeaf'
-export * from './RollupState'
 export * from './VaultLeaf'
